requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.88.0
fastapi-utils==0.2.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
pycryptodome==3.17
pydantic==1.10.7
python-telegram-bot==20.7
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.22.0
tinytuya==1.9.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.20.0
